[HERDR] Smart Cleanup v0.05: integer age signals, lenient decode, public logs
Root cause of the work-Mac "stall": the backend emitted age signals as raw
floats (sessionFileAgeSeconds: 369398.1364490986) while Swift's
CleanupSignals typed them Int? — JSONDecoder throws "not representable" and
the whole envelope decode fails, so every poll of a finished report errors
and the sheet freezes on the last in-flight frame. Bites exactly when panes
have pi session files or done alerts, i.e. always on a real machine.

- Backend: _age_int() coerces all *AgeSeconds signals to int(round()).
- Swift: those fields decode as Double? so a stray float can never kill the
  envelope again (regression fixture with the exact failing values).
- os_log interpolations marked privacy .public — no more "<private>" in
  `log show` output.
- CleanupFeature.version -> v0.05.

Verified by decoding the captured failing payload with the app's own model
files: DECODE FAILED before, DECODE OK after.

+++ herdr-harness-mac/herdr-harness-mac/Models/CleanupReport.swift
@@ -397,9 +397,9 @@ enum CleanupClassification: String, Codable, Sendable, Equatable {
 }
 
 struct CleanupSignals: Codable, Sendable, Equatable {
-    let doneAlertAgeSeconds: Int?
+    let doneAlertAgeSeconds: Double?
     let revisionChanged: Bool?
-    let sessionFileAgeSeconds: Int?
+    let sessionFileAgeSeconds: Double?
     let starred: Bool?
     let focused: Bool?
     let unreadAlerts: Int?
